# docs/swagger.yaml
basePath: /api
definitions:
  account.GetAccountResponse:
    properties:
      balance:
        type: integer
      id:
        type: integer
      name:
        type: string
      type:
        type: string
    type: object
  handler.createAccountRequestBody:
    properties:
      name:
        type: string
    type: object
  handler.createUserRequestBody:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  handler.deleteAccountRequestBody:
    properties:
      name:
        type: string
    type: object
  handler.makeTransferRequestBody:
    properties:
      amount:
        type: integer
      payee_account_id:
        type: integer
      sender_account_id:
        type: integer
    type: object
  handler.userLoginRequestBody:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API Server for Mock Banking Application
  title: Mock Banking Application API
  version: "1.0"
paths:
  /account:
    delete:
      consumes:
      - application/json
      description: Delete a bank account with specific name
      operationId: deletes custom bank account
      parameters:
      - description: account name
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.deleteAccountRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Delete a bank account
      tags:
      - Bank Account
    post:
      consumes:
      - application/json
      description: Create a new custom bank account with specific name
      operationId: creates custom bank account
      parameters:
      - description: account name
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.createAccountRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Create a bank account
      tags:
      - Bank Account
  /accounts:
    get:
      description: Get all bank accounts you have
      operationId: gets all bank accounts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/account.GetAccountResponse'
            type: array
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get all bank accounts
      tags:
      - Bank Account
  /deposits:
    get:
      description: Get all deposits other clients have made to your accounts
      operationId: Gets deposits
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get all deposits
      tags:
      - Deposit
  /login:
    post:
      consumes:
      - application/json
      description: Login into account
      operationId: logins into account
      parameters:
      - description: user data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.userLoginRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Login
      tags:
      - Auth
  /logout:
    post:
      description: Logout from account
      operationId: logs out from account
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: Logout
      tags:
      - Auth
  /register:
    post:
      consumes:
      - application/json
      description: Register new user
      operationId: Registers new user
      parameters:
      - description: user data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.createUserRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Register
      tags:
      - Auth
  /transfer:
    post:
      consumes:
      - application/json
      description: Make a transfer between your/other clients accounts
      operationId: Makes a transfer
      parameters:
      - description: transfer data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.makeTransferRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema: {}
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Make a transfer
      tags:
      - Transfer
  /transfers:
    get:
      description: Get all transfers you have made
      operationId: Gets transfers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get all transfers
      tags:
      - Transfer
swagger: "2.0"